api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "FastAPI backend for Recycling Buddy"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.42.38",
    "fastapi>=0.128.0",
    "pillow>=12.1.0",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.22",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "httpx>=0.28.1",
    "pyrefly>=0.19.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.14",
]

[tool.pyrefly]
search_path = ["."]

[tool.pyright]
# Override automatic src-layout detection
# Set import root to api directory (not api/src)
# This allows imports like 'from src.main import app' to work
include = ["src", "tests"]
stubPath = ""
venvPath = "."
venv = ".venv"
